// src/modules/textile/reports/knitting/party-wise-knitting-program-report/party-wise-knitting-program-report-type.ts
export type IPartyWiseKnittingProgram = {
    DETAILS_ID: number;
    KNITTING_PROGRAM_NO: string;
    KNITTING_PROG_DATE: Date;
    COMPANY_ID: number;
    COMPANY_NAME: string;
    COMPANY_ADDRESS: string;
    MAIN_COMPANY_NAME: string;
    MAIN_COMPANY_ADDRESS: string;
    KNITTING_HOUSE_ID: number;
    KNIT_HOUSE: string;
    KNIT_HOUSE_ADDRESS: string;
    BUYERID: number;
    BUYER: string;
    PO_ID: number;
    PONO: string;
    STYLE_ID: number;
    STYLENO: string;
    YARN: string;
    YARN_LOT: string;
    BRAND_NAME: string;
    FABRIC: string;
    MC_DIA: string;
    FINISH_DIA: string;
    GSM: string;
    GMT_COLOR_ID: number;
    COLORNAME: string;
    YCOLOR: string;
    STITCH_LENGTH: string;
    FABRIC_ID: number;
    GAUGE: string;
    KNITTING_RATE: number;
    START_DATE: string | null;
    END_DATE: string | null;
    REMARKS: string;
    LYCRA_CM: string;
    FABRIC_PART: string;
    DTLS_QTY: number;
    DTLS_PICES: number;
    DTLS_TOTAL_KNIT_AMOUNT: number;
    KNITTING_ADVICE: string;
    PREPARED_BY: string;
    FABRIC_TYPE: string;
    SIZENAME: string;
    ORDER_NO: string;
    IS_CONSIDER_AS_RIB_FOR_REPORT: string;
    KNIT_HOUSE_CODE: string;
    YARN_SUPPLIER_CODE: string;
};